Atualiza estrutura do projeto e integra código do backend e mobile atualizados

+++ back/src/main/java/br/com/eletrodescarte/models/PontoColeta.java
@@ -7,7 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -31,8 +30,8 @@ public class PontoColeta {
     @Column(nullable = false)
     private String endereco;
 
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private String telefone;
     private String email;
     private boolean ativo;
